CHEF-2601 : Review comments
